camera_model/requirements.txt
numpy>=1.24
Pillow>=10.0
matplotlib>=3.7
streamlit>=1.32
